third_party/blink/web_tests/platform/generic/http/tests/security/isolatedWorld/didClearWindowObject-expected.txt

PASS: __worldID is undefined in normal world.
FAIL: __worldID is undefined in isolated world 1.
